Share global data across all frontend schemas
This adds the global field to all frontend schemas so that they can have
header and footer data from the homepage schema embedded into them.

As part of this I've moved the global field to be a definition and
extracted the link property fields to have their own definition. To
avoid collision with the existing loaded word of links I've given these
the super 90s term hyperlinks. Suggestions of better names are welcome.