Add new examples for direct attachment rendering
Previously we relied on pre-rendered HTML for the attached 'documents'
of publications and consultations. This adds an example to support work
to directly render attachments based on metadata in the payload, and an
array of featured attachment IDs. See RFC 116 for more details.

Note that 'documents' is a required field for publications, so we still
need to populate it with an empty array.

https://github.com/alphagov/govuk-rfcs/blob/master/rfc-116-store-attachment-data-in-content-items.md