Move phase message to be defined in details body
This should be set-up for all the formats mainstream publisher can
publish this field too.

It's out of the root of the schema as specifying a message there
requires adding additional fields to Publishing and Content Store.
Arguably the phase field should all be defined in details as it by no
means ubiquitous across GOV.UK.