Add 'ordering' to Take Part pages in schema
Take Part pages save an order in the model in Whitehall so that
they can be displayed in a specific order on the "Get Involved"
page. However this ordering was not being exported to Publishing
API, limiting our ability to migrate away from Whitehall.

Storing the ordering in Publishing API allows us to migrate the
frontend parts away from Whitehall, as this key information would
now be in a centralised place.

Trello: https://trello.com/c/aAW8oqCk/2174-5add-ordering-to-take-part-pages-in-schema