Add body to step by steps
Site search can only ingest the title and description of a step by step
at present. The body of a step by step is written in markdown-ish, which
we can convert to html using kramdown or similar. By adding this html to
the body field of the content item, we'll be able to search for them much
more easily.