Add smart_answer schema
Previously, content items used to respresent an Smart Answer were a
transaction schema for the start page and a
generic_with_external_related_links for the flow. This is because start
pages were rendered by Frontend. However, we are changing Smart Answers
to render the start pages and represent flows with a single content
item. This adds schema to use for that content item.