Deprecate supertypes that aren't used downstream
These supertypes used to be consumed by downstream apps (search-api and
content-data). We've done some work to remove this dependency, so that the
only app still left relying on finding supertypes via a notification is
email-alert-api.

Our strategy is that supergroups are a point in time feature of a document
based on its type. We should therefore devolve working these out to apps
that wish to use them, and minimise the complexity of everything else.

- [Content-data-api not longer looks for them](alphagov/content-data-api@48a5c97#diff-cbf8b25d6d853acf58fa9057841f407d29ffe90649c75cf9e3f51b2d6e3aa1d3)
- [Search-api works them out for itself](https://github.com/alphagov/search-api/blob/987fa2313c0b35528eb87b892dccbd1cfd692dca/lib/indexer/document_preparer.rb#L155)
- [Publishing-api still sends them, but we don't want it to](https://github.com/alphagov/publishing-api/blob/a6ec79b0a8d77b991e0dd192f89dd88f9fe0b687/app/presenters/edition_presenter.rb#L52)

Email alert api still uses government_document_supertype and email_document_supertype
so we'll leave them for another day.

https://trello.com/c/hp5BJD2i/117-unused-incorrect-supertypes-in-content-items