Devolved administration availability schema for local transactions
This schema has been added as part of a task to enable publishers to
specify devolved administration service availability for local transactions.

In the Frontend application, by default, users will be shown a link to a
local authority website when inputting a postcode for a local transaction.

In some cases, the service is handled by a devolved nation. In other cases
the service is unavailable to devolved nations.

Currently, the devolved nation availability cannot be set by publishers and
the logic is defined in the frontend application.

As part of the overall task, functionality is being added to the Publisher
application to allow this information to be set within the publishing tool.

Therefore, this schema validates the DA availability values set in Publisher.

Only the 'unavailable' and 'devolved_administration_service' related values
will be presented from Publisher. This will be enough information to
determine what to display on the Frontend as it will fall back to displaying
the local authority results page.

[Devolved Administration Availability Data modelling PR in Publisher](alphagov/publisher#1456)
Co-authored-by: Kevin Dew <kevin.dew@digital.cabinet-office.gov.uk>