Fix nested list markdown
The automatic numbering would reset at the 'once the tests pass' <li>,
probably due to the mixture of ordered and unordered list types in this
list. There's no reason the point above it needs to be in a nested list,
so I've moved it all onto the same (rendered) line.