Simplify the URI required for redirect destinations
This changes the schema validation for redirect destinations from
something quite restrictive to something much more permissive. The
reason for taking a permissive approach here is because of duplication
of validation efforts.

Currently Publishing API has an involved redirect destination validation
process, which is used for both unpublishing and individual redirects [1].
This schema has not kept up with this validation and is currently out of
sync. Rather than trying to continue a process of keeping a schema and a
Rails validator in sync, it seems simpler and more accurate to make the
schema permissive and rely on the Publishing API validator as the
validating entity.

The prompt for this change was trying to adjust Short URL Manager to
allow redirects to hosts off GOV.UK (nhs.uk) this had the pain point
that Short URL Manager validates the URL, then Publishing API validates
the URL via Content Schema, and then Publishing API validates the URL
again through its Rails validator. The distribution of this logic is a
pain point.

[1]: https://github.com/alphagov/publishing-api/blob/3717c4057c0c05f9846a122ceb5972e559117d7c/app/validators/routes_and_redirects_validator.rb#L122