Add Account API to the list of rendering apps
The Account API can be used to personalise pages for logged in users.
Following [Account API ADR 2][1] this is intended to approached as a
progressive enhancement via javascript in a browser.

As such these endpoints can now be called and will return payloads of
JSON, sometimes containing rendered HTML to be added into application
frontends.

As we wish to proxy these routes to account-api we will need to identify
account-api a rendering app.

[1]: https://github.com/alphagov/account-api/blob/main/docs/adr/002-progressive-enhancement.md