Replace mainstream_browse_type with mainstream_browse_origin in Topic
We want to reference the Content id of Mainstream browse page the topic
originated from as a part of unified topic system work.

Until the work is completed we will need to keep the Mainstream browse page and
the cloned Topic is sync. Since other attributes of the Mainstream browse page
such as title or slug can change, this adds a unique referenece we can use to
identify the cloned Topic.

The presence of this attribute can still acts as a flag.