remove placeholder topical event document type
This was replaced by a non placeholder document type, and there are 0 content items in the content store with this type now.

There were previously two examples using the 'placeholder_topical_event' schema and document type. I've removed one of these since it seems unecessary to test both, and moved the other to the correct folder to reflect the non-placeholder schema and document type.